requirements.txt
asgiref==3.6.0
attrs==22.2.0
certifi==2022.12.7
distlib==0.3.6
dj-database-url==1.2.0
Django==4.1.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.25.1
filelock==3.9.0
inflection==0.5.1
jsonschema==4.17.3
pipenv==2023.2.4
platformdirs==3.0.0
psycopg2==2.9.5
PyJWT==2.6.0
pyrsistent==0.19.3
pytz==2022.7.1
PyYAML==6.0
sqlparse==0.4.3
uritemplate==4.1.1
virtualenv==20.19.0
virtualenv-clone==0.5.7
